# pkg/api/api.openapi.yaml
openapi: "3.0.4"
info:
  title: Finances API
  version: "1.0"
servers:
  - url: https://api.server.test/v1
paths:
  /books:
    get:
      operationId: GetBooks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 200
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                required:
                  - books
                properties:
                  books:
                    type: array
                    items:
                      $ref: "model.openapi.yaml#/components/schemas/Book"
    post:
      operationId: CreateBook
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              required:
                - book
              properties:
                book:
                  $ref: "model.openapi.yaml#/components/schemas/BookProperties"
      responses:
        "201":
          description: Created
          content:
            "application/json":
              schema:
                type: object
                required:
                  - book
                properties:
                  book:
                    $ref: "model.openapi.yaml#/components/schemas/Book"
  "/books/{bookId}":
    parameters:
      - name: bookId
        in: path
        schema:
          type: string
    get:
      operationId: GetBook
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                properties:
                  book:
                    $ref: "model.openapi.yaml#/components/schemas/Book"
        "404":
          description: Not Found
  /exchanges:
    get:
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                required:
                  - exchanges
                properties:
                  exchanges:
                    type: array
                    items:
                      $ref: "model.openapi.yaml#/components/schemas/ExchangeWithSplits"
  /health:
    get:
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: object
                properties:
                  status:
                    $ref: "model.openapi.yaml#/components/schemas/ServerHealthStatus"
                  reason:
                    type: string
                required:
                  - status